dummy_data/fake_data_service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
faker>=19.0.0
pandas>=2.0.0
reportlab>=4.0.0
pyyaml>=6.0.0
python-multipart>=0.0.6
